+++ Amnesia/Data/Config.cs
@@ -14,9 +14,9 @@ namespace Amnesia.Data {
         public static string WarnAtLifeName { get; private set; } = "WarnAtLife";
         public static string EnablePositiveOutlookName { get; private set; } = "EnablePositiveOutlook";
         public static string ForgetLevelsAndSkillsName { get; private set; } = "ForgetLevelsAndSkills";
-        public static string ForgetActiveQuests { get; private set; } = "ForgetActiveQuests";
-        public static string ForgetInactiveQuests { get; private set; } = "ForgetInactiveQuests";
-        public static string ForgetIntroQuests { get; private set; } = "ForgetIntroQuests";
+        public static string ForgetActiveQuestsName { get; private set; } = "ForgetActiveQuests";
+        public static string ForgetInactiveQuestsName { get; private set; } = "ForgetInactiveQuests";
+        public static string ForgetIntroQuestsName { get; private set; } = "ForgetIntroQuests";
 
         private static readonly string disconnectionWarning = "\n        - [!] SYSTEM WILL DISCONNECT PLAYER ON FINAL DEATH IF ENABLED!";
         private static readonly Dictionary<string, string> FieldNamesAndDescriptionsDict = new Dictionary<string, string> {
@@ -24,9 +24,9 @@ namespace Amnesia.Data {
             { WarnAtLifeName, "number of lives remaining when system should start warning players about amnesia" },
             { EnablePositiveOutlookName, $"whether to grant temporary buff that boosts xp growth at initial server join and on memory loss" },
             { ForgetLevelsAndSkillsName, "whether to player levels and skills should be forgotten on memory loss" },
-            { ForgetActiveQuests, $"whether ongoing quests should be forgotten on memory loss{disconnectionWarning}" },
-            { ForgetInactiveQuests, $"whether completed quests (AND TRADER TIER LEVELS) should be forgotten on memory loss{disconnectionWarning}" },
-            { ForgetIntroQuests, $"whether the intro quests should be forgotten/reset on memory loss{disconnectionWarning}" }
+            { ForgetActiveQuestsName, $"whether ongoing quests should be forgotten on memory loss{disconnectionWarning}" },
+            { ForgetInactiveQuestsName, $"whether completed quests (AND TRADER TIER LEVELS) should be forgotten on memory loss{disconnectionWarning}" },
+            { ForgetIntroQuestsName, $"whether the intro quests should be forgotten/reset on memory loss{disconnectionWarning}" }
         };
         public static List<string> FieldNames { get; private set; } = FieldNamesAndDescriptionsDict.Keys.ToList();
         public static string FieldNamesAndDescriptions { get; private set; } = "    - " + string.Join("\n    - ", FieldNamesAndDescriptionsDict.Select(kvp => kvp.Key + ": " + kvp.Value));
@@ -34,26 +34,21 @@ namespace Amnesia.Data {
         public static int MaxLives { get; private set; } = 2;
         public static int WarnAtLife { get; private set; } = 1;
         public static bool EnablePositiveOutlook { get; private set; } = true;
-        // TODO: add/support options in console for this
-        public static bool ResetLevels { get; private set; } = true;
+        public static bool ForgetLevelsAndSkills { get; private set; } = true;
 
-        // TODO: recommend against this + clearIntroQuests both being enabled if not also resetting levels
-        // TODO: default to false
-        // TODO: add/support options in console for this
-        public static bool ResetQuests { get; private set; } = false;
-        // TODO: add/support options in console for this
-        public static bool ClearIntroQuests { get; private set; } = false;
-        public static bool ResetFactionPoints { get; private set; } = false;
+        public static bool ForgetActiveQuests { get; private set; } = false;
+        public static bool ForgetInactiveQuests { get; private set; } = false;
+        public static bool ForgetIntroQuests { get; private set; } = false;
 
         public static string AsString() {
             return $@"=== Amnesia Configuration ===
 {MaxLivesName}: {MaxLives}
 {WarnAtLifeName}: {WarnAtLife}
 {EnablePositiveOutlookName}: {EnablePositiveOutlook}
-{ForgetLevelsAndSkillsName}: {ResetLevels}
-{ForgetActiveQuests}: {ResetQuests}
-{ForgetInactiveQuests}: {ResetFactionPoints}
-{ForgetIntroQuests}: {ClearIntroQuests}";
+{ForgetLevelsAndSkillsName}: {ForgetLevelsAndSkills}
+{ForgetActiveQuestsName}: {ForgetActiveQuests}
+{ForgetInactiveQuestsName}: {ForgetInactiveQuests}
+{ForgetIntroQuestsName}: {ForgetIntroQuests}";
         }
 
         /**
@@ -112,8 +107,8 @@ namespace Amnesia.Data {
          * <param name="value">New value to use.</param>
          */
         public static void SetResetLevels(bool value) {
-            if (ResetLevels != value) {
-                ResetLevels = value;
+            if (ForgetLevelsAndSkills != value) {
+                ForgetLevelsAndSkills = value;
                 Save();
             }
         }
@@ -123,8 +118,8 @@ namespace Amnesia.Data {
          * <param name="value">New value to use.</param>
          */
         public static void SetResetQuests(bool value) {
-            if (ResetQuests != value) {
-                ResetQuests = value;
+            if (ForgetActiveQuests != value) {
+                ForgetActiveQuests = value;
                 Save();
             }
         }
@@ -134,8 +129,8 @@ namespace Amnesia.Data {
          * <param name="value">New value to use.</param>
          */
         public static void SetClearIntroQuests(bool value) {
-            if (ClearIntroQuests != value) {
-                ClearIntroQuests = value;
+            if (ForgetIntroQuests != value) {
+                ForgetIntroQuests = value;
                 Save();
             }
         }
@@ -145,8 +140,8 @@ namespace Amnesia.Data {
          * <param name="value">New value to use.</param>
          */
         public static void SetResetFactionPoints(bool value) {
-            if (ResetFactionPoints != value) {
-                ResetFactionPoints = value;
+            if (ForgetInactiveQuests != value) {
+                ForgetInactiveQuests = value;
                 Save();
             }
         }
@@ -157,10 +152,10 @@ namespace Amnesia.Data {
                     new XElement(MaxLivesName, MaxLives),
                     new XElement(WarnAtLifeName, WarnAtLife),
                     new XElement(EnablePositiveOutlookName, EnablePositiveOutlook),
-                    new XElement(ForgetLevelsAndSkillsName, ResetLevels),
-                    new XElement(ForgetActiveQuests, ResetQuests),
-                    new XElement(ForgetIntroQuests, ClearIntroQuests),
-                    new XElement(ForgetInactiveQuests, ResetFactionPoints)
+                    new XElement(ForgetLevelsAndSkillsName, ForgetLevelsAndSkills),
+                    new XElement(ForgetActiveQuestsName, ForgetActiveQuests),
+                    new XElement(ForgetIntroQuestsName, ForgetIntroQuests),
+                    new XElement(ForgetInactiveQuestsName, ForgetInactiveQuests)
                 ).Save(filename);
                 log.Info($"Successfully saved {filename}");
                 return true;
@@ -176,10 +171,10 @@ namespace Amnesia.Data {
                 MaxLives = ParseInt(config, MaxLivesName);
                 WarnAtLife = ParseInt(config, WarnAtLifeName);
                 EnablePositiveOutlook = ParseBool(config, EnablePositiveOutlookName);
-                ResetLevels = ParseBool(config, ForgetLevelsAndSkillsName);
-                ResetQuests = ParseBool(config, ForgetActiveQuests);
-                ClearIntroQuests = ParseBool(config, ForgetIntroQuests);
-                ResetFactionPoints = ParseBool(config, ForgetInactiveQuests);
+                ForgetLevelsAndSkills = ParseBool(config, ForgetLevelsAndSkillsName);
+                ForgetActiveQuests = ParseBool(config, ForgetActiveQuestsName);
+                ForgetIntroQuests = ParseBool(config, ForgetIntroQuestsName);
+                ForgetInactiveQuests = ParseBool(config, ForgetInactiveQuestsName);
                 log.Info($"Successfully loaded {filename}");
                 return true;
             } catch (FileNotFoundException) {